pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-model-core"
version = "0.1.0"
description = "Core face embedding training and inference pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
